Several updates to the function:

- More robust output
- Enhanced error handling, skip AMIs that aren't found
- Updated AMI tags for the completed and revoked events

+++ function/lambda_function.py
@@ -1,13 +1,11 @@
 # ---
-import boto3
+import boto3, botocore.exceptions
 import hashlib
 import hmac
 import json
 import os
 import urllib3
 from datetime import datetime, timezone, timedelta
-
-# TODO: Handle InvalidAMIID.NotFound exceptions
 
 
 def lambda_handler(event, context):
@@ -18,10 +16,7 @@ def lambda_handler(event, context):
     try:
         signature = event["headers"]["X-Hcp-Webhook-Signature"]
     except KeyError:
-        return {
-            "statusCode": 403,
-            "body": "HMAC signature not provided.",
-        }
+        return {"statusCode": 403, "body": "HMAC signature not provided."}
 
     # --- Verify the HMAC, then check the eventAction value to determine what to do.
     if verify_hmac(event):
@@ -41,117 +36,185 @@ def lambda_handler(event, context):
             case _:
                 return {
                     "statusCode": 400,
-                    "body": f'Action {body["eventAction"]} found in request is not supported.',
+                    "body": f"Action {body['eventAction']} found in request is not supported.",
                 }
     else:
-        return {
-            "statusCode": 403,
-            "body": "Unauthorized: HMAC signature mismatch.",
-        }
+        return {"statusCode": 403, "body": "Unauthorized: HMAC signature mismatch."}
 
 
 # --- Event Action functions
 def verify():
     """Handle the HCP Packer webhook verification event."""
-    response = {
-        "statusCode": "200",
-        "body": "Verification successful!",
-    }
+    response = {"statusCode": "200", "body": "Verification successful!"}
     return response
 
 
 def complete(body):
     """Handle the HCP Packer 'Completed iteration' webhook event. Adds metadata tags to the AMI(s)."""
+    result = {"actions": []}
     try:
         amis = return_image_id(body["eventPayload"]["builds"], "aws")
+        if len(amis) == 0:
+            return {"statusCode": 200, "body": "No AMIs found in iteration."}
 
         for ami in amis:
             ec2_client = boto3.client("ec2", region_name=ami["region"])
+            result_status = "Success"
+            result_message = "AMI tags added"
             ami_id = ami["id"]
 
             # Create a tag for the AMI with the HCP Packer metadata
-            ec2_client.create_tags(
-                Resources=[ami_id],
-                Tags=[
-                    {
-                        "Key": "HCPPackerBucket",
-                        "Value": body["eventPayload"]["bucket"]["slug"],
-                    },
-                    {
-                        "Key": "HCPPackerIterationFingerprint",
-                        "Value": body["eventPayload"]["iteration"]["fingerprint"],
-                    },
-                    {
-                        "Key": "HCPPackerBuildID",
-                        "Value": ami["build_id"],
-                    },
-                ],
+            try:
+                ec2_client.create_tags(
+                    Resources=[ami_id],
+                    Tags=[
+                        {
+                            "Key": "HCPPackerBucket",
+                            "Value": body["eventPayload"]["bucket"]["slug"],
+                        },
+                        {
+                            "Key": "HCPPackerIterationFingerprint",
+                            "Value": body["eventPayload"]["iteration"]["fingerprint"],
+                        },
+                        {
+                            "Key": "HCPPackerIterationVersion",
+                            "Value": body["eventPayload"]["iteration"]["version"],
+                        },
+                        {
+                            "Key": "HCPPackerBuildID",
+                            "Value": ami["build_id"],
+                        },
+                    ],
+                )
+
+            except botocore.exceptions.ClientError as e:
+                error_code = e.response["Error"]["Code"]
+                if error_code.startswith("InvalidAMIID"):
+                    result_status = "Skipped"
+                    result_message = error_code
+                else:
+                    raise e
+
+            result["actions"].append(
+                {
+                    "ami_id": ami_id,
+                    "region": ami["region"],
+                    "status": result_status,
+                    "message": result_message,
+                }
             )
 
-        return {
-            "statusCode": 200,
-            "body": json.dumps(f"Successfully added AMI metadata tags."),
-        }
+        return {"statusCode": 200, "body": json.dumps(result)}
 
     except Exception as e:
-        return {"statusCode": 500, "body": json.dumps(f"Error: {str(e)}")}
+        return {"statusCode": 500, "body": f"Error: {str(e)}"}
 
 
 def revoke(body):
     """Handle the HCP Packer 'Revoked iteration' webhook event. Sets the AMI deprecation time and adds metadata tags."""
     # Set the deprecation time to the current time plus 1 minute (EnableImageDeprecation won't accept a time in the past)
     deprecation_time = datetime.now(timezone.utc) + timedelta(minutes=1)
-    deprecation_reason = body["eventPayload"]["iteration"]["revocation_message"]
+    result = {"actions": []}
 
     try:
         # The revoke webhook doesn't include the builds object, so we need to call get_builds to get them
         amis = return_image_id(get_builds(body), "aws")
+        if len(amis) == 0:
+            return {"statusCode": 200, "body": "No AMIs found in iteration."}
 
         for ami in amis:
             ec2_client = boto3.client("ec2", region_name=ami["region"])
+            result_status = "Success"
+            result_message = f"AMI deprecation time set to {deprecation_time}"
             ami_id = ami["id"]
 
-            # Enable deprecation for the AMI
-            ec2_client.enable_image_deprecation(
-                ImageId=ami_id, DeprecateAt=deprecation_time, DryRun=False
+            try:
+                # Enable deprecation for the AMI
+                ec2_client.enable_image_deprecation(
+                    ImageId=ami_id, DeprecateAt=deprecation_time, DryRun=False
+                )
+
+                # Create a tag for the AMI with the deprecation time
+                ec2_client.create_tags(
+                    Resources=[ami_id],
+                    Tags=[
+                        {
+                            "Key": "HCPPackerRevoked",
+                            "Value": "true",
+                        },
+                        {
+                            "Key": "HCPPackerRevokedBy",
+                            "Value": body["eventPayload"]["iteration"][
+                                "revocation_author"
+                            ],
+                        },
+                        {
+                            "Key": "HCPPackerRevocationMessage",
+                            "Value": body["eventPayload"]["iteration"][
+                                "revocation_message"
+                            ],
+                        },
+                    ],
+                )
+
+            except botocore.exceptions.ClientError as e:
+                error_code = e.response["Error"]["Code"]
+                if error_code.startswith("InvalidAMIID"):
+                    result_status = "Skipped"
+                    result_message = error_code
+                else:
+                    raise e
+
+            result["actions"].append(
+                {
+                    "ami_id": ami_id,
+                    "region": ami["region"],
+                    "status": result_status,
+                    "message": result_message,
+                }
             )
 
-            # Create a tag for the AMI with the deprecation time
-            ec2_client.create_tags(
-                Resources=[ami_id],
-                Tags=[
-                    {"Key": "DeprecationTime", "Value": deprecation_time.isoformat()},
-                    {
-                        "Key": "DeprecationReason",
-                        "Value": deprecation_reason,
-                    },
-                ],
-            )
-
-        return {
-            "statusCode": 200,
-            "body": json.dumps(
-                f"Successfully set deprecation time to {deprecation_time}."
-            ),
-        }
+        return {"statusCode": 200, "body": json.dumps(result)}
 
     except Exception as e:
-        return {"statusCode": 500, "body": json.dumps(f"Error: {str(e)}")}
+        return {"statusCode": 500, "body": f"Error: {str(e)}"}
 
 
 def delete(body):
     """Handle the HCP Packer 'Deleted iteration' webhook event. Deregisters the AMI(s) and deletes their associated snapshots."""
+    result = {"actions": []}
     try:
         amis = return_image_id(body["eventPayload"]["builds"], "aws")
+        if len(amis) == 0:
+            return {"statusCode": 200, "body": "No AMIs found in iteration."}
 
         for ami in amis:
             ec2_client = boto3.client("ec2", region_name=ami["region"])
+            result_status = "Success"
+            result_message = "AMI deregistered"
             ami_id = ami["id"]
 
             # Deregister the AMI
-            ec2_client.deregister_image(ImageId=ami_id)
+            try:
+                ec2_client.deregister_image(ImageId=ami_id)
+            except botocore.exceptions.ClientError as e:
+                error_code = e.response["Error"]["Code"]
+                if error_code.startswith("InvalidAMIID"):
+                    result_status = "Skipped"
+                    result_message = error_code
+                else:
+                    raise e
 
-            # Describe the snapshots associated with the AMI
+            result["actions"].append(
+                {
+                    "ami_id": ami_id,
+                    "region": ami["region"],
+                    "status": result_status,
+                    "message": result_message,
+                }
+            )
+
+            # Get the snapshots associated with the AMI
             snapshots = ec2_client.describe_snapshots(
                 Filters=[{"Name": "description", "Values": [f"*{ami_id}*"]}]
             )["Snapshots"]
@@ -159,44 +222,72 @@ def delete(body):
             # Delete the associated snapshots
             for snapshot in snapshots:
                 ec2_client.delete_snapshot(SnapshotId=snapshot["SnapshotId"])
+                result["actions"].append(
+                    {
+                        "snapshot_id": snapshot["SnapshotId"],
+                        "region": ami["region"],
+                        "status": "Success",
+                        "message": f"Snapshot for AMI {ami_id} deleted",
+                    }
+                )
 
-        return {
-            "statusCode": 200,
-            "body": json.dumps(
-                f"Successfully deregistered AMI(s) and deleted associated snapshot(s)."
-            ),
-        }
+        return {"statusCode": 200, "body": json.dumps(result)}
 
     except Exception as e:
-        return {"statusCode": 500, "body": json.dumps(f"Error: {str(e)}")}
+        return {"statusCode": 500, "body": f"Error: {str(e)}"}
 
 
 def restore(body):
     """Handle the HCP Packer 'Restored iteration' webhook event. Disables deprecation on the AMI and removes the tag added by the 'revoke' handler."""
+    result = {"actions": []}
     try:
         # The restore webhook doesn't include the builds object, so we need to call get_builds to get them
         amis = return_image_id(get_builds(body), "aws")
+        if len(amis) == 0:
+            return {"statusCode": 200, "body": "No AMIs found in iteration."}
 
         for ami in amis:
             ec2_client = boto3.client("ec2", region_name=ami["region"])
+            result_status = "Success"
+            result_message = "AMI deprecation cleared"
+            ami_id = ami["id"]
 
-            # Disable deprecation for the AMI
-            ec2_client.disable_image_deprecation(ImageId=ami["id"], DryRun=False)
+            try:
+                # Disable deprecation for the AMI
+                ec2_client.disable_image_deprecation(ImageId=ami_id, DryRun=False)
 
-            # Delete the DeprecationTime tag
-            ec2_client.delete_tags(
-                DryRun=False,
-                Resources=[ami["id"]],
-                Tags=[{"Key": "DeprecationTime"}, {"Key": "DeprecationReason"}],
+                # Delete the DeprecationTime tag
+                ec2_client.delete_tags(
+                    DryRun=False,
+                    Resources=[ami_id],
+                    Tags=[
+                        {"Key": "HCPPackerRevoked"},
+                        {"Key": "HCPPackerRevokedBy"},
+                        {"Key": "HCPPackerRevocationMessage"},
+                    ],
+                )
+
+            except botocore.exceptions.ClientError as e:
+                error_code = e.response["Error"]["Code"]
+                if error_code.startswith("InvalidAMIID"):
+                    result_status = "Skipped"
+                    result_message = error_code
+                else:
+                    raise e
+
+            result["actions"].append(
+                {
+                    "ami_id": ami_id,
+                    "region": ami["region"],
+                    "status": result_status,
+                    "message": result_message,
+                }
             )
 
-        return {
-            "statusCode": 200,
-            "body": json.dumps(f"AMI deprecation disabled."),
-        }
+        return {"statusCode": 200, "body": json.dumps(result)}
 
     except Exception as e:
-        return {"statusCode": 500, "body": json.dumps(f"Error: {str(e)}")}
+        return {"statusCode": 500, "body": f"Error: {str(e)}"}
 
 
 # --- Helper functions
